File organization and misc updates
Re-numbered the scripts for more clarity around the order in which they are to be run.
Updated ssmsproj file, which was referencing nonexistent files
Consolidated some of the logic, reducing the total number of files
Move the import.People script to the end, pointing out in the comments that it is not integrated with the rest of the data.